+++ FreeAIr/MCP/Agent/Github/GithubMCPInstallBackgroundTask.cs
@@ -15,6 +15,7 @@ namespace FreeAIr.MCP.Agent.Github
         public GithubMCPInstallBackgroundTask(
             )
         {
+            StartAsyncTask();
         }
 
         protected override async Task RunWorkingTaskAsync()