+++ FreeAIr/UI/Windows/ControlCenterWindow.xaml.cs
@@ -1,0 +1,17 @@
+﻿using System.Threading;
+using System.Windows;
+using System.Windows.Input;
+
+namespace FreeAIr.UI.Windows
+{
+    public partial class ControlCenterWindow : Window
+    {
+        public ControlCenterWindow(
+            )
+        {
+            InitializeComponent();
+        }
+
+    }
+
+}