+++ RELEASE_NOTES.md
@@ -6,6 +6,11 @@ Please report any bugs to the [github repo](https://github.com/lsoft/FreeAIr).
 
 If you are enjoying FreeAIr to the enough level to donate, there are a [lot of small cancer patients](https://advitausa.org/au/index.php/donate/) that need your help.
 
+## 0.2.4
+
+- Added a FreeAIr Control Center window.
+- Added a global MCP tools status, and chat MCP tools status.
+
 ## 0.2.3
 
 - Added a natural language search (across whole solution or current project).